requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# litGL/fontDistiller.py: 26
# litGL/fontDistiller_cmap.py: 26
Pillow >= 8.2.0

# litGL/app.py: 16
# litGL/examples/example.py: 3
# litGL/examples/example_1.py: 16
# litGL/examples/example_2.py: 3
# litGL/examples/example_3.py: 16
# litGL/examples/example_4.py: 17
# litGL/examples/example_5.py: 17
# litGL/examples/example_6.py: 17
# litGL/examples/glfwBackend.py: 7
# litGL/litGL.py: 18
# litGL/shader.py: 10
# litGL/transform.py: 8
PyGLM >= 1.99.3

# litGL/app.py: 8
# litGL/examples/example.py: 2
# litGL/examples/example_1.py: 8
# litGL/examples/example_2.py: 2
# litGL/examples/example_3.py: 8
# litGL/examples/example_4.py: 8
# litGL/examples/example_5.py: 8
# litGL/examples/example_6.py: 7
# litGL/examples/glfwBackend.py: 6
# litGL/font.py: 2
# litGL/litGL.py: 15
# litGL/mesh.py: 2
# litGL/shader.py: 8
# litGL/texture.py: 2
PyOpenGL >= 3.1.5

# litGL/fontDistiller.py: 33
# litGL/fontDistiller_cmap.py: 33
fonttools >= 4.18.2

# litGL/examples/example.py: 4
# litGL/examples/example_2.py: 4
# litGL/examples/example_4.py: 10
# litGL/examples/example_5.py: 10
# litGL/examples/example_6.py: 8
# litGL/examples/glfwBackend.py: 4
glfw >= 2.0.0

# litGL/app.py: 7
# litGL/examples/example.py: 1
# litGL/examples/example_1.py: 7
# litGL/examples/example_2.py: 1
# litGL/examples/example_3.py: 7
# litGL/examples/example_4.py: 7
# litGL/examples/example_5.py: 7
# litGL/examples/example_6.py: 6
# litGL/examples/glfwBackend.py: 3
# litGL/font.py: 1
# litGL/fontDistiller.py: 35,36
# litGL/fontDistiller_cmap.py: 35,36
# litGL/litGL.py: 16
# litGL/mesh.py: 1
# litGL/shader.py: 9
# litGL/texture.py: 1
numpy >= 1.17.4
